Refactor package structure and add packaging files
