+++ fastapi_toolkit/__init__.py
@@ -1,0 +1,2 @@
+__all__ = ['asynchronous', 'exceptions', 'schema', 'servicios']
+__version__ = '0.1.0'
