+++ fastapi_toolkit/asynchronous/service/__init__.py
@@ -1,0 +1,1 @@
+from .base import BaseService
