+++ fastapi_toolkit/exceptions/__init__.py
@@ -1,0 +1,2 @@
+from .api_exceptions import *
+from .handler import *
