+++ fastapi_toolkit/servicios/thrid/__init__.py
@@ -1,0 +1,2 @@
+from .jwt import JWTService
+from .supabase import SupabaseService, get_supabase_service
